LayoutTests/fast/canvas/webgl/webgl-clear-composited-notshowing-expected.txt

Test that test that WebGL canvas preserveDrawingBuffer = false clear behavior on cases where compositor does not neccessarily display the element. You should see no failures and some of the canvases as green on screen.
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS successfullyParsed is true

TEST COMPLETE
